agents/monitor_a2a/runtime/requirements.txt
strands-agents>=1.10.0
strands-agents-tools
bedrock-agentcore>=1.4.0
a2a-sdk>=0.3.24,<1.0
fastapi>=0.115.0
uvicorn>=0.30.0
